bb.fetch.git: add support for shallow mirror tarballs
This adds support to the git fetcher for fetching and generating mirror
tarballs of shallow git repositories, with the specified revisions / depths.
The external git-make-shallow script is used for shallow mirror tarball
creation.

This implements support for shallow mirror tarballs, not shallow clones.
Supporting shallow clones directly would be rather more problematic, as we'd
need to hardcode the depth between branch HEAD and the SRCREV, and that depth
would change as the branch is updated. Sadly, git's remote protocol is
extremely limited.

When BB_GIT_SHALLOW is enabled, we will always attempt to fetch a shallow
mirror tarball. If the shallow mirror tarball cannot be fetched, it will try
to fetch the full mirror tarball and use that.

Revisions to be removed from the repository can be specified by branch+depth
or by explicitly listing them.

The shallow construction process scales with the number of refs to be
processed, which means a repository with a ton of tags, for example, greatly
slows the process. As a result, the default behavior is to remove all unused
refs from the repository except those referenced by the URL and those
explicitly specified in BB_GIT_SHALLOW_EXTRA_REFS.

- BB_GIT_SHALLOW: enable/disable shallow support as a whole, boolean
- BB_GIT_SHALLOW_DEPTH: specify commit depth for the included refs. defaults
  to 1 (the top commit alone). This can be set to 0 or the empty string to
  disable.
- BB_GIT_SHALLOW_DEPTH_name: same as above, for a specific named url/branch
- BB_GIT_SHALLOW_REVS: specific revisions whose history should be removed,
  beyond that specified by refs+depth
- BB_GIT_SHALLOW_REVS_name: same as above, for a specific named url/branch
- BB_GIT_SHALLOW_EXTRA_REFS: additional refs to keep beyond those referenced
  in SRC_URI, supports wildcards. I.e. `refs/heads/*`
- BB_GENERATE_SHALLOW_TARBALLS: explicitly control creation of shallow mirror
  tarballs. this defaults to enabled when BB_GENERATE_MIRROR_TARBALLS is
  enabled

Example usage:

    BB_GIT_SHALLOW ?= "1"
    BB_GIT_SHALLOW_DEPTH ?= "1"

    # Usage in a recipe with multiple named uris or multiple named branches
    BB_GIT_SHALLOW_DEPTH_doc = ""

    # Remove the upstream history from our kernel repo, keeping our own
    BB_GIT_SHALLOW_REVS_pn-linux-mel_mx6 = "v3.14"
    BB_GIT_SHALLOW_DEPTH_pn-linux-mel_mx6 = "0"

    # Keep the branches in linux-yocto repositories
    BB_GIT_SHALLOW_EXTRA_REFS_pn-linux-yocto = "refs/heads/*"

Signed-off-by: Christopher Larson <chris_larson@mentor.com>