bb.fetch: support multiple mirror tarball filenames
If a fetcher lists multiple mirror tarball filenames in ud.mirrortarballs, it
will attempt to fetch and use each of them, stopping at the first it
successfully fetches. This is needed for git shallow mirror tarball support,
as we want to be able to use either a shallow or full mirror tarball.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>