bb.fetch: don't remove the clone when an update fails
If our clone is out of date, and the attempt to update it fails, we want to
fall back to mirror tarballs as appropriate, but we don't necessarily want to
have to re-clone if no tarball is available. So don't blindly delete the git
repository if an update fails, remove it when extracting a full mirror
tarball, and adjust the git method logic to account for the fact that an out
of date clonedir may exist.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>