bb.fetch: don't remove the clone when an update fails
When our clone exists, but is out of date, and the attempt to update it fails,
we don't necessarily want to remove the entire clone, particularly if it's
a large repository. Instead, keep it around, and if we have a mirror tarball
when downloading, remove the clonedir and extract the tarball at that time.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>