git-make-shallow: add script to make a git repo shallow
This script will be used by the git fetcher to create shallow mirror tarballs.

The specified revisions will have their history removed from the current git
repository, across all refs, or the specified refs. If refs are specified,
only those refs are kept, and all others are removed.

Git's shallow support is implemented using grafts. Any commits listed in
.git/shallow get their parents removed -- grafted to nothing. If we only
remove the parents of a specified commit, i.e. 5 commits deep in the current
branch, it's still possible for the history to be kept. For example, a commit
after the specified commit could be a merge of a commit older than the
specified commit, effectively bypassing the graft by pulling in the history
via a different path in the commit graph. If this occurs, the repository is
effectively not shallow, so it's not enough to just add the specific revs we
want to remove, or the commit revs at a particular depth.

We need to graft at every point where the history is pulled in. `git fetch
--depth` and `git clone --depth` handle this, but it's not fine grained enough
for our usage, as this depth is either for a single branch, or all branches,
and only works for branch+depth, not a specified revision.

To avoid this, git-make-shallow has a more involved process. For the specified
revisions, across all or specified refs, we use git merge-base to identify any
points where the history of the revisions intersects the history of the refs,
then we also make those points shallow, and continue the process walking into
the history until we run out of points where the history intersects. This
produces results similar to what git does for its more limited option, but is
more flexible.

As the work done, and time taken, scales with the number of refs, we filter
down the refs to just the ones we care about, as is usually the case for `git
clone --depth`. Even a `git fetch --depth=` can take a massive amount of time
in a linux kernel repository, for example, so this is something we want to
avoid. We also truncate branches to SRCREV to further reduce the repository
size (that is, commits made after SRCREV but before branch HEAD are removed).

Signed-off-by: Christopher Larson <chris_larson@mentor.com>