git-make-shallow: add script to make a git repo shallow
This script will be used by the git fetcher to create shallow mirror tarballs.

The specified revisions will have their history removed from the current git
repository, across all refs, or the specified refs. If refs are specified,
only those refs are kept, and all others are removed.

Git's shallow support is implemented using grafts. Any commits listed in
.git/shallow get their parents removed -- grafted to nothing. If we only
remove the parents of a specified commit, i.e. 5 commits deep in the current
branch, it's still possible for the history to be kept. For example, a commit
after the specified commit could be a merge of a commit older than the
specified commit, effectively bypassing the graft by pulling in the history
via a different path in the commit graph. If this occurs, the repository is
effectively not shallow, so this is insufficient to our needs.

To avoid this, git-make-shallow has a more involved process. For the specified
revisions, across all or specified refs, we use git merge-base to identify any
points where the history of the revisions intersects the history of the refs,
then we also make those points shallow, and continue the process walking into
the history until we run out of points where the history intersects. This
doesn't scale well with larger numbers of refs, which is why use of an
explicit list of refs is recommended.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>