bb.cooker: add 'bblayers_only' argument to BBCooker
When set, only the bblayers.conf + layer.conf files are parsed, not
bitbake.conf. Any use of updateCache() will result in re-parsing all the base
configuration files in such a case, to ensure the config is parsed, otherwise
if the user wants to switch from bblayers_only to just parsing the config,
they can re-run initConfigurationData().

Signed-off-by: Christopher Larson <chris_larson@mentor.com>