bb.tests.fetch: add tests for shallow
These test shallow via the bitbake fetcher rather than the underlying script
it uses to do the work, so cover the behavior of the API.

To be added at a later date: test to ensure that everything needed,
configuration wise, is encoded into the mirror tarball filename. I.e. a change
to the branches we're including needs to result in a change in the filename,
otherwise we could pick up an invalid tarball.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>